refactor(credential): split AcceptedSet/builder and dedupe key reconciliation
Separate the AcceptedSet type from its builder into accepted_set.go and
accepted_set_builder.go, and split the tests to match (builder tests, error
type test, and Rotator reconcile tests each in their own file).

Extract the shared add/remove diff in reconcileSDKKeys and reconcileMobileKeys
into a generic reconcileAcceptedKeys helper so the algorithm lives in one
place, and annotate each pass so the diff logic reads clearly.

Drop the stale ordering comment on primaryCredentials.
